backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
requests
ffmpeg-python
python-multipart
